+++ src/Ext/SWType/Body.cpp
@@ -45,6 +45,7 @@ void SWTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->SW_Next_RollChances)
 		.Process(this->ShowTimer_Priority)
 		.Process(this->Convert_Pairs)
+		.Process(this->Convert_Anim)
 		;
 }
 
@@ -165,6 +166,7 @@ void SWTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	Nullable<AffectedHouse> convertAffectedHouses;
 	convertFrom.Read(exINI, pSection, "Convert.From");
 	convertTo.Read(exINI, pSection, "Convert.To");
+	Convert_Anim.Read(exINI, pSection, "Convert.Anim");
 	convertAffectedHouses.Read(exINI, pSection, "Convert.AffectedHouses");
 	if (convertTo.isset())
 	{