+++ src/Ext/WarheadType/Detonate.cpp
@@ -338,7 +338,7 @@ void WarheadTypeExt::ExtData::ApplyConvert(HouseClass* pHouse, TechnoClass* pTar
 	if (!pTargetFoot || this->Convert_Pairs.size() == 0)
 		return;
 
-	TypeConvertHelper::Convert(pTargetFoot, this->Convert_Pairs, pHouse);
+	TypeConvertHelper::Convert(pTargetFoot, this->Convert_Pairs, pHouse, this->Convert_Anim);
 }
 
 void WarheadTypeExt::ExtData::ApplyLocomotorInfliction(TechnoClass* pTarget)