+++ src/Misc/TypeConvertHelper.h
@@ -2,10 +2,11 @@
 
 #include <Utilities/EnumFunctions.h>
 #include <Ext/Techno/Body.h>
+#include <Ext/AnimType/Body.h>
 
 namespace TypeConvertHelper
 {
 	typedef std::vector<std::tuple<ValueableVector<TechnoTypeClass*>, NullableIdx<TechnoTypeClass>, Nullable<AffectedHouse>>> ConvertPairs;
 
-	void Convert(FootClass* pTargetFoot, ConvertPairs& convertPairs, HouseClass* pOwner);
+	void Convert(FootClass* pTargetFoot, ConvertPairs& convertPairs, HouseClass* pOwner, AnimTypeClass* pTypeAnim);
 }